+++ src/parser.rs
@@ -1,769 +1,19 @@
-use crate::ast::{Block, Document, Inline, ListItem, ListKind, SourceMap};
-use crate::error::{ErrorHandler, ErrorInfo, ErrorSeverity, MarkdownError, Result};
-use crate::lexer::{Lexer, Position, token as nom_token};
-// use crate::lexer::{Lexer, Position, token::Token};
-use crate::performance::ParallelConfig;
-
-/// Configuration for the parser behavior and options
-#[derive(Debug, Clone)]
-pub struct ParserConfig {
-    /// Whether to strictly follow CommonMark specification
-    pub strict_commonmark: bool,
-    /// Maximum nesting depth to prevent stack overflow
-    pub max_nesting_depth: usize,
-    /// Whether to enable GitHub Flavored Markdown extensions
-    pub enable_gfm_extensions: bool,
-    /// Whether to collect detailed source position information
-    pub track_source_positions: bool,
-    /// Maximum number of errors before stopping parsing
-    pub max_errors: Option<usize>,
-    /// Whether to enable performance optimizations
-    pub performance_optimized: bool,
-    /// Parallel processing configuration
-    pub parallel_config: ParallelConfig,
-}
-
-impl Default for ParserConfig {
-    fn default() -> Self {
-        Self {
-            strict_commonmark: true,
-            max_nesting_depth: 64,
-            enable_gfm_extensions: false,
-            track_source_positions: true,
-            max_errors: Some(100),
-            performance_optimized: true,
-            parallel_config: ParallelConfig::default(),
-        }
-    }
-}
-
-/// Enhanced nom-based parser supporting comprehensive CommonMark constructs.
-///
-/// This parser leverages the nom-based lexer to provide robust parsing of:
-/// - Block elements: headings, paragraphs, code blocks, blockquotes, lists, thematic breaks
-/// - Inline elements: text, emphasis, code spans, links, images, HTML inlines
-/// - Advanced features: error recovery, position tracking, reference definitions
-pub struct Parser<'input> {
-    lexer: Lexer<'input>,
-    config: ParserConfig,
-    current_token: Option<nom_token::Token<'input>>,
-    peeked_token: Option<nom_token::Token<'input>>,
-    error_handler: Box<dyn ErrorHandler>,
-    reference_definitions: std::collections::HashMap<String, crate::ast::LinkReference>,
-}
-
-impl<'input> Parser<'input> {
-    /// Creates a new parser instance with the given configuration.
-    pub fn new(input: &'input str, config: ParserConfig) -> Self {
-        let mut lexer = Lexer::new(input);
-        let current_token = lexer.next_token();
-
-        let error_handler: Box<dyn ErrorHandler> = if let Some(max_errors) = config.max_errors {
-            Box::new(crate::error::DefaultErrorHandler::with_max_errors(
-                max_errors,
-            ))
-        } else {
-            Box::new(crate::error::DefaultErrorHandler::new())
-        };
-
-        Self {
-            lexer,
-            config,
-            current_token,
-            peeked_token: None,
-            error_handler,
-            reference_definitions: std::collections::HashMap::new(),
-        }
-    }
-
-    /// Convenience helper that uses default parser configuration.
-    pub fn with_defaults(input: &'input str) -> Self {
-        Self::new(input, ParserConfig::default())
-    }
-
-    /// Creates a parser with a custom error handler.
-    pub fn with_error_handler(
-        input: &'input str,
-        config: ParserConfig,
-        error_handler: Box<dyn ErrorHandler>,
-    ) -> Self {
-        let mut lexer = Lexer::new(input);
-        let current_token = lexer.next_token();
-
-        Self {
-            lexer,
-            config,
-            current_token,
-            peeked_token: None,
-            error_handler,
-            reference_definitions: std::collections::HashMap::new(),
-        }
-    }
-
-    /// Returns the current position in the input.
-    pub fn position(&self) -> Position {
-        self.lexer.position()
-    }
-
-    /// Returns a reference to the parser configuration.
-    pub fn config(&self) -> &ParserConfig {
-        &self.config
-    }
-
-    /// Returns a reference to the collected reference definitions.
-    pub fn reference_definitions(
-        &self,
-    ) -> &std::collections::HashMap<String, crate::ast::LinkReference> {
-        &self.reference_definitions
-    }
-
-    /// Parses the input into a `Document` AST.
-    pub fn parse(mut self) -> Result<Document> {
-        let mut blocks = Vec::new();
-        let source_map = SourceMap::new();
-
-        // Parse blocks until we reach EOF
-        while !self.is_at_end() {
-            // Skip empty lines
-            if matches!(self.current_token, Some(nom_token::Token::LineEnding(_))) {
-                self.advance()?;
-                continue;
-            }
-
-            // Skip indentation at document level
-            if matches!(self.current_token, Some(nom_token::Token::Indent(_))) {
-                self.advance()?;
-                continue;
-            }
-
-            // Try to parse link reference definitions first
-            if self.try_parse_link_reference_definition()? {
-                continue;
-            }
-
-            // Parse the next block
-            match self.parse_block() {
-                Ok(block) => blocks.push(block),
-                Err(error) => {
-                    self.handle_parse_error(error.clone());
-                    if self.config.strict_commonmark {
-                        return Err(error);
-                    }
-                    // In non-strict mode, try to recover
-                    self.attempt_recovery();
-                }
-            }
-        }
-
-        Ok(Document { blocks, source_map })
-    }
-
-    /// Advances to the next token.
-    fn advance(&mut self) -> Result<()> {
-        if let Some(peeked) = self.peeked_token.take() {
-            self.current_token = Some(peeked);
-        } else {
-            self.current_token = self.lexer.next_token();
-        }
-        Ok(())
-    }
-
-    /// Peeks at the next token without consuming it.
-    fn peek_token(&mut self) -> Option<&nom_token::Token<'input>> {
-        if self.peeked_token.is_none() {
-            self.peeked_token = self.lexer.next_token();
-        }
-        self.peeked_token.as_ref()
-    }
-
-    /// Checks if we've reached the end of input.
-    fn is_at_end(&self) -> bool {
-        matches!(self.current_token, Some(nom_token::Token::Eof) | None)
-    }
-
-    /// Attempts error recovery by advancing to a safe parsing state.
-    fn attempt_recovery(&mut self) {
-        // Try to find a block boundary or other safe recovery point
-        while !self.is_at_end() {
-            match &self.current_token {
-                Some(nom_token::Token::LineEnding(_)) | Some(nom_token::Token::Eof) => {
-                    self.advance().ok();
-                    break;
-                }
-                Some(nom_token::Token::AtxHeading(_))
-                | Some(nom_token::Token::SetextHeading(_))
-                | Some(nom_token::Token::ThematicBreak(_))
-                | Some(nom_token::Token::BlockQuote(_)) => {
-                    // Found a block-level element, good recovery point
-                    break;
-                }
-                _ => {
-                    if self.advance().is_err() {
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    /// Handles parse errors with appropriate recovery.
-    fn handle_parse_error(&mut self, error: MarkdownError) {
-        let error_info = ErrorInfo::new(
-            if error.is_recoverable() {
-                ErrorSeverity::Error
-            } else {
-                ErrorSeverity::Fatal
-            },
-            format!("Parse error: {}", error),
-        )
-        .with_position(self.position());
-
-        self.error_handler.handle_error(&error_info);
-
-        if error.is_recoverable() && !self.config.strict_commonmark {
-            self.attempt_recovery();
-        }
-    }
-
-    /// Parses a single block element.
-    fn parse_block(&mut self) -> Result<Block> {
-        // Clone the token to avoid borrowing issues
-        let current_token = self.current_token.clone();
-
-        match current_token {
-            Some(nom_token::Token::AtxHeading(heading)) => self.parse_atx_heading(&heading),
-            Some(nom_token::Token::SetextHeading(heading)) => self.parse_setext_heading(&heading),
-            Some(nom_token::Token::CodeBlock(code_block)) => self.parse_code_block(&code_block),
-            Some(nom_token::Token::BlockQuote(blockquote)) => self.parse_blockquote(&blockquote),
-            Some(nom_token::Token::ListMarker(list_marker)) => self.parse_list(&list_marker),
-            Some(nom_token::Token::ThematicBreak(_)) => self.parse_thematic_break(),
-            Some(nom_token::Token::HtmlBlock(html_block)) => self.parse_html_block(&html_block),
-            _ => {
-                // Default to paragraph parsing for text and other inline content
-                self.parse_paragraph()
-            }
-        }
-    }
-
-    /// Parses an ATX heading.
-    fn parse_atx_heading(&mut self, heading: &nom_token::AtxHeadingToken<'input>) -> Result<Block> {
-        let position = Some(self.position());
-        let level = heading.level;
-        let content_text = heading.raw_content.trim();
-
-        self.advance()?; // consume the heading token
-
-        // Parse the heading content as inline elements
-        let inline_content = self.parse_text_as_inlines(content_text)?;
-
-        Ok(Block::Heading {
-            level,
-            content: inline_content,
-            id: None,
-            position,
-        })
-    }
-
-    /// Parses a Setext heading.
-    fn parse_setext_heading(
-        &mut self,
-        heading: &nom_token::SetextHeadingToken<'input>,
-    ) -> Result<Block> {
-        let position = Some(self.position());
-        let level = heading.level;
-
-        self.advance()?; // consume the setext heading token
-
-        // For Setext headings, we need the text from the previous line
-        // This is a simplified implementation - in practice, we'd need to track previous content
-        let inline_content = vec![Inline::Text("Setext Heading".to_string())];
-
-        Ok(Block::Heading {
-            level,
-            content: inline_content,
-            id: None,
-            position,
-        })
-    }
-
-    /// Parses a code block.
-    fn parse_code_block(
-        &mut self,
-        code_block: &nom_token::CodeBlockToken<'input>,
-    ) -> Result<Block> {
-        let position = Some(self.position());
-
-        self.advance()?; // consume the code block token
-
-        // Extract language from info string if present
-        let language = code_block
-            .info_string
-            .and_then(|info| info.split_whitespace().next().map(|s| s.to_string()));
-
-        Ok(Block::CodeBlock {
-            info: code_block.info_string.map(|s| s.to_string()),
-            content: code_block.raw_content.to_string(),
-            language,
-            position,
-        })
-    }
-
-    /// Parses a blockquote.
-    fn parse_blockquote(&mut self, _blockquote: &nom_token::BlockQuoteToken) -> Result<Block> {
-        let position = Some(self.position());
-
-        self.advance()?; // consume the blockquote marker
-
-        let mut content = Vec::new();
-
-        // Parse the content of the blockquote
-        while !self.is_at_end() {
-            // Skip whitespace after blockquote marker
-            if matches!(self.current_token, Some(nom_token::Token::Indent(_))) {
-                self.advance()?;
-                continue;
-            }
-
-            // Check if we've reached the end or a non-blockquote line
-            if matches!(self.current_token, Some(nom_token::Token::LineEnding(_))) {
-                self.advance()?;
-                // Check if the next line continues the blockquote
-                if matches!(self.current_token, Some(nom_token::Token::BlockQuote(_))) {
-                    continue;
-                } else {
-                    // End of blockquote
-                    break;
-                }
-            }
-
-            // If we hit another blockquote marker, continue parsing
-            if matches!(self.current_token, Some(nom_token::Token::BlockQuote(_))) {
-                self.advance()?;
-                continue;
-            }
-
-            // Parse a block within the blockquote
-            match self.parse_block() {
-                Ok(block) => content.push(block),
-                Err(_) => break, // End of blockquote content
-            }
-        }
-
-        Ok(Block::BlockQuote { content, position })
-    }
-
-    /// Parses a list.
-    fn parse_list(&mut self, first_marker: &nom_token::ListMarkerToken<'input>) -> Result<Block> {
-        let position = Some(self.position());
-        let mut items = Vec::new();
-        let mut tight = true;
-
-        // Convert nom token ListKind to AST ListKind
-        let list_kind = match &first_marker.marker {
-            nom_token::ListKind::Bullet { marker } => ListKind::Bullet { marker: *marker },
-            nom_token::ListKind::Ordered { start, delimiter } => ListKind::Ordered {
-                start: *start,
-                delimiter: *delimiter,
-            },
-        };
-
-        // Parse list items
-        loop {
-            let should_continue = match &self.current_token {
-                Some(nom_token::Token::ListMarker(marker)) => {
-                    self.list_markers_match(&first_marker.marker, &marker.marker)
-                }
-                _ => false,
-            };
-
-            if !should_continue {
-                break;
-            }
-
-            let item = self.parse_list_item()?;
-            items.push(item);
-
-            // Check for blank lines between items (makes list loose)
-            if self.has_blank_line_before_next_item() {
-                tight = false;
-            }
-        }
-
-        Ok(Block::List {
-            kind: list_kind,
-            tight,
-            items,
-            position,
-        })
-    }
-
-    /// Parses a single list item.
-    fn parse_list_item(&mut self) -> Result<ListItem> {
-        self.advance()?; // consume the list marker
-
-        let mut content = Vec::new();
-        let tight = true;
-
-        // Skip whitespace after list marker
-        while matches!(self.current_token, Some(nom_token::Token::Whitespace(_))) {
-            self.advance()?;
-        }
-
-        // Parse the first line of the list item
-        if !matches!(
-            self.current_token,
-            Some(nom_token::Token::LineEnding(_)) | Some(nom_token::Token::Eof)
-        ) {
-            let block = self.parse_block()?;
-            content.push(block);
-        }
-
-        // Handle continuation lines
-        while matches!(self.current_token, Some(nom_token::Token::LineEnding(_))) {
-            self.advance()?; // consume newline
-
-            // Check if next line is indented (continuation of list item)
-            if matches!(self.current_token, Some(nom_token::Token::Indent(_))) {
-                self.advance()?; // consume indentation
-                if !self.is_at_end()
-                    && !matches!(self.current_token, Some(nom_token::Token::ListMarker(_)))
-                {
-                    let block = self.parse_block()?;
-                    content.push(block);
-                }
-            } else {
-                // End of list item
-                break;
-            }
-        }
-
-        Ok(ListItem {
-            content,
-            tight,
-            task_list_marker: None, // Task list support can be added later
-        })
-    }
-
-    /// Parses a thematic break.
-    fn parse_thematic_break(&mut self) -> Result<Block> {
-        let position = Some(self.position());
-
-        self.advance()?; // consume the thematic break token
-
-        Ok(Block::ThematicBreak { position })
-    }
-
-    /// Parses an HTML block.
-    fn parse_html_block(
-        &mut self,
-        html_block: &nom_token::HtmlBlockToken<'input>,
-    ) -> Result<Block> {
-        let position = Some(self.position());
-
-        self.advance()?; // consume the HTML block token
-
-        Ok(Block::HtmlBlock {
-            content: html_block.raw.to_string(),
-            position,
-        })
-    }
-
-    /// Parses a paragraph.
-    fn parse_paragraph(&mut self) -> Result<Block> {
-        let position = Some(self.position());
-        let mut inline_content = Vec::new();
-
-        // Collect inline content until we hit a paragraph boundary
-        while !self.is_at_end() && !self.is_paragraph_boundary() {
-            // Clone the current token to avoid borrowing issues
-            let current_token = self.current_token.clone();
-
-            match current_token {
-                Some(nom_token::Token::Text(text_token)) => {
-                    let inlines = self.parse_text_as_inlines(text_token.lexeme)?;
-                    inline_content.extend(inlines);
-                    self.advance()?;
-                }
-                Some(nom_token::Token::Whitespace(ws)) => {
-                    inline_content.push(Inline::Text(ws.lexeme.to_string()));
-                    self.advance()?;
-                }
-                Some(nom_token::Token::EmphasisDelimiter(emphasis)) => {
-                    let emphasis_inline = self.parse_emphasis_delimiter(&emphasis)?;
-                    inline_content.push(emphasis_inline);
-                }
-                Some(nom_token::Token::CodeSpan(code_span)) => {
-                    inline_content.push(Inline::Code(code_span.content.to_string()));
-                    self.advance()?;
-                }
-                Some(nom_token::Token::Autolink(autolink)) => {
-                    let link_inline = self.parse_autolink(&autolink)?;
-                    inline_content.push(link_inline);
-                }
-                Some(nom_token::Token::Entity(entity)) => {
-                    let text_inline = self.parse_entity(&entity)?;
-                    inline_content.push(text_inline);
-                }
-                Some(nom_token::Token::EscapeSequence(escape)) => {
-                    inline_content.push(Inline::Text(escape.escaped.to_string()));
-                    self.advance()?;
-                }
-                Some(nom_token::Token::HtmlInline(html)) => {
-                    inline_content.push(Inline::HtmlInline(html.raw.to_string()));
-                    self.advance()?;
-                }
-                Some(nom_token::Token::LineEnding(_)) => {
-                    // Check if this is the end of the paragraph
-                    self.advance()?;
-                    if self.is_paragraph_boundary() {
-                        break;
-                    }
-                    // Otherwise, add a soft break and continue
-                    inline_content.push(Inline::SoftBreak);
-                }
-                _ => {
-                    // Hit a block-level token; ensure we always make progress.
-                    if inline_content.is_empty() {
-                        // Advance once to avoid getting stuck on the same token.
-                        self.advance()?;
-                    }
-                    break;
-                }
-            }
-        }
-
-        // If no content was collected, create a paragraph with empty text
-        if inline_content.is_empty() {
-            inline_content.push(Inline::Text(String::new()));
-        }
-
-        Ok(Block::Paragraph {
-            content: inline_content,
-            position,
-        })
-    }
-
-    /// Parses text content as inline elements.
-    fn parse_text_as_inlines(&mut self, text: &str) -> Result<Vec<Inline>> {
-        // For now, treat as simple text
-        // In a full implementation, this would parse nested inline elements
-        if text.is_empty() {
-            Ok(vec![Inline::Text(String::new())])
-        } else {
-            Ok(vec![Inline::Text(text.to_string())])
-        }
-    }
-
-    /// Parses an emphasis delimiter into an emphasis inline.
-    fn parse_emphasis_delimiter(
-        &mut self,
-        emphasis: &nom_token::EmphasisDelimiterToken,
-    ) -> Result<Inline> {
-        let strong = emphasis.run_length >= 2;
-
-        self.advance()?; // consume opening delimiter
-
-        let mut content = Vec::new();
-
-        // Look for closing delimiter
-        while !self.is_at_end() {
-            match &self.current_token {
-                Some(nom_token::Token::Text(text_token)) => {
-                    content.push(Inline::Text(text_token.lexeme.to_string()));
-                    self.advance()?;
-                }
-                Some(nom_token::Token::EmphasisDelimiter(closing_emphasis))
-                    if closing_emphasis.marker == emphasis.marker
-                        && closing_emphasis.run_length >= emphasis.run_length =>
-                {
-                    self.advance()?; // consume closing delimiter
-                    break;
-                }
-                Some(nom_token::Token::LineEnding(_)) => {
-                    // Emphasis can span soft breaks but not hard breaks
-                    content.push(Inline::SoftBreak);
-                    self.advance()?;
-                }
-                _ => {
-                    // Hit a block boundary, end emphasis
-                    break;
-                }
-            }
-        }
-
-        Ok(Inline::Emphasis { strong, content })
-    }
-
-    /// Parses an autolink into a link inline.
-    fn parse_autolink(&mut self, autolink: &nom_token::AutolinkToken<'input>) -> Result<Inline> {
-        self.advance()?; // consume autolink token
-
-        let text = vec![Inline::Text(autolink.lexeme.to_string())];
-        let destination = autolink.lexeme.to_string();
-
-        Ok(Inline::Link {
-            text,
-            destination,
-            title: None,
-        })
-    }
-
-    /// Parses an entity into a text inline.
-    fn parse_entity(&mut self, entity: &nom_token::EntityToken<'input>) -> Result<Inline> {
-        self.advance()?; // consume entity token
-
-        // In a full implementation, this would resolve the entity to a character
-        let resolved_text = entity
-            .resolved
-            .map(|c| c.to_string())
-            .unwrap_or_else(|| format!("&{}", entity.raw));
-
-        Ok(Inline::Text(resolved_text))
-    }
-
-    /// Checks if the current position represents a paragraph boundary.
-    fn is_paragraph_boundary(&self) -> bool {
-        matches!(
-            &self.current_token,
-            Some(nom_token::Token::AtxHeading(_))
-                | Some(nom_token::Token::SetextHeading(_))
-                | Some(nom_token::Token::CodeBlock(_))
-                | Some(nom_token::Token::BlockQuote(_))
-                | Some(nom_token::Token::ListMarker(_))
-                | Some(nom_token::Token::ThematicBreak(_))
-                | Some(nom_token::Token::HtmlBlock(_))
-                | Some(nom_token::Token::Eof)
-                | None
-        )
-    }
-
-    /// Checks if two list markers match (same type).
-    fn list_markers_match(
-        &self,
-        marker1: &nom_token::ListKind,
-        marker2: &nom_token::ListKind,
-    ) -> bool {
-        matches!(
-            (marker1, marker2),
-            (
-                nom_token::ListKind::Bullet { .. },
-                nom_token::ListKind::Bullet { .. }
-            ) | (
-                nom_token::ListKind::Ordered { .. },
-                nom_token::ListKind::Ordered { .. }
-            )
-        )
-    }
-
-    /// Checks for blank lines between list items.
-    fn has_blank_line_before_next_item(&mut self) -> bool {
-        // Look ahead to see if there's a blank line
-        if let Some(nom_token::Token::LineEnding(_)) = self.peek_token() {
-            // Check if the token after the line ending is another line ending (blank line)
-            true // Simplified - would need more sophisticated lookahead
-        } else {
-            false
-        }
-    }
-
-    /// Tries to parse a link reference definition.
-    fn try_parse_link_reference_definition(&mut self) -> Result<bool> {
-        // Link reference definitions are complex to parse from tokens
-        // This is a simplified placeholder implementation
-        // In practice, this would need to recognize the pattern and parse it properly
-        Ok(false)
-    }
-}
-
-/// Enhanced paragraph builder that handles complex inline content.
-#[cfg(test)]
-struct ParagraphBuilder {
-    inlines: Vec<Inline>,
-    consecutive_newlines: usize,
-    has_content: bool,
-}
+// Parser module with submodules for better organization using Rust 2018 module system
+mod config;
+mod core;
+mod utils;
 
 #[cfg(test)]
-impl ParagraphBuilder {
-    fn new() -> Self {
-        Self {
-            inlines: Vec::new(),
-            consecutive_newlines: 0,
-            has_content: false,
-        }
-    }
+mod tests;
 
-    fn push_inline(&mut self, inline: Inline) {
-        // Skip empty text inlines
-        if let Inline::Text(text) = &inline
-            && text.trim().is_empty()
-            && !self.has_content
-        {
-            return;
-        }
+// Re-export main types and functions
+pub use config::ParserConfig;
+pub use core::Parser;
+pub use utils::ParagraphBuilder;
+// NomParser is now the main Parser
 
-        self.inlines.push(inline);
-        self.consecutive_newlines = 0;
-        self.has_content = true;
-    }
-
-    fn push_text(&mut self, fragment: &str) {
-        if fragment.trim().is_empty() && !self.has_content {
-            return;
-        }
-
-        // Merge with previous text inline if possible
-        if let Some(Inline::Text(last_text)) = self.inlines.last_mut() {
-            if !last_text.ends_with(' ') && !fragment.starts_with(' ') {
-                last_text.push(' ');
-            }
-            last_text.push_str(fragment.trim_end_matches('\n'));
-        } else {
-            self.inlines
-                .push(Inline::Text(fragment.trim_end_matches('\n').to_string()));
-        }
-
-        self.consecutive_newlines = 0;
-        self.has_content = true;
-    }
-
-    fn flush_into(&mut self, output: &mut Vec<Block>) {
-        if let Some(block) = self.flush() {
-            output.push(block);
-        }
-    }
-
-    fn flush(&mut self) -> Option<Block> {
-        if !self.has_content || self.inlines.is_empty() {
-            self.clear();
-            return None;
-        }
-
-        // Remove trailing soft breaks
-        while let Some(Inline::SoftBreak) = self.inlines.last() {
-            self.inlines.pop();
-        }
-
-        if self.inlines.is_empty() {
-            self.clear();
-            return None;
-        }
-
-        let content = std::mem::take(&mut self.inlines);
-        self.clear();
-
-        Some(Block::Paragraph {
-            content,
-            position: None,
-        })
-    }
-
-    fn clear(&mut self) {
-        self.inlines.clear();
-        self.consecutive_newlines = 0;
-        self.has_content = false;
-    }
-}
+use crate::ast::Document;
+use crate::error::{ErrorHandler, Result};
 
 /// Parse Markdown using the default parser configuration.
 pub fn parse(markdown: &str) -> Result<Document> {
@@ -784,423 +34,4 @@ pub fn parse_with_error_handler(
 ) -> Result<Document> {
     let parser = Parser::with_error_handler(markdown, config, error_handler);
     parser.parse()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parses_heading_and_paragraph() {
-        let input = "# Heading\n\nSome text here.";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("nom parser should succeed");
-
-        assert_eq!(doc.blocks.len(), 2);
-        match &doc.blocks[0] {
-            Block::Heading { level, .. } => assert_eq!(*level, 1),
-            _ => panic!("expected heading"),
-        }
-        match &doc.blocks[1] {
-            Block::Paragraph { content, .. } => {
-                assert_eq!(content.len(), 1);
-                assert!(matches!(&content[0], Inline::Text(text) if text == "Some text here."));
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn parses_multiple_headings() {
-        let input = "# Level 1\n## Level 2\n### Level 3";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse multiple headings");
-
-        assert_eq!(doc.blocks.len(), 3);
-
-        for (i, expected_level) in [1u8, 2u8, 3u8].iter().enumerate() {
-            match &doc.blocks[i] {
-                Block::Heading { level, .. } => assert_eq!(*level, *expected_level),
-                _ => panic!("expected heading at index {}", i),
-            }
-        }
-    }
-
-    #[test]
-    fn parses_code_blocks() {
-        let input = "```rust\nfn main() {}\n```";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse code block");
-
-        assert_eq!(doc.blocks.len(), 1);
-        match &doc.blocks[0] {
-            Block::CodeBlock {
-                info,
-                content,
-                language,
-                ..
-            } => {
-                assert_eq!(info.as_deref(), Some("rust"));
-                assert_eq!(language.as_deref(), Some("rust"));
-                assert!(content.contains("fn main()"));
-            }
-            _ => panic!("expected code block"),
-        }
-    }
-
-    #[test]
-    fn parses_blockquotes() {
-        let input = "> This is a blockquote\n> Second line";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse blockquote");
-
-        assert_eq!(doc.blocks.len(), 1);
-        match &doc.blocks[0] {
-            Block::BlockQuote { content, .. } => {
-                assert!(!content.is_empty());
-            }
-            _ => panic!("expected blockquote"),
-        }
-    }
-
-    #[test]
-    fn parses_bullet_lists() {
-        let input = "- Item 1\n- Item 2\n- Item 3";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse bullet list");
-
-        assert_eq!(doc.blocks.len(), 1);
-        match &doc.blocks[0] {
-            Block::List { kind, items, .. } => {
-                assert!(matches!(kind, ListKind::Bullet { .. }));
-                assert_eq!(items.len(), 3);
-            }
-            _ => panic!("expected list"),
-        }
-    }
-
-    #[test]
-    fn parses_ordered_lists() {
-        let input = "1. First item\n2. Second item\n3. Third item";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse ordered list");
-
-        assert_eq!(doc.blocks.len(), 1);
-        match &doc.blocks[0] {
-            Block::List { kind, items, .. } => {
-                assert!(matches!(kind, ListKind::Ordered { start: 1, .. }));
-                assert_eq!(items.len(), 3);
-            }
-            _ => panic!("expected ordered list"),
-        }
-    }
-
-    #[test]
-    fn parses_thematic_breaks() {
-        let input = "---\n\n***\n\n___";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse thematic breaks");
-
-        let thematic_break_count = doc
-            .blocks
-            .iter()
-            .filter(|block| matches!(block, Block::ThematicBreak { .. }))
-            .count();
-
-        assert!(
-            thematic_break_count >= 1,
-            "should parse at least one thematic break"
-        );
-    }
-
-    #[test]
-    fn parses_mixed_content() {
-        let input = "# Heading\n\nParagraph text.\n\n- List item\n\n```\ncode\n```";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse mixed content");
-
-        assert!(doc.blocks.len() >= 4);
-
-        // Check that we have different block types
-        let has_heading = doc
-            .blocks
-            .iter()
-            .any(|b| matches!(b, Block::Heading { .. }));
-        let has_paragraph = doc
-            .blocks
-            .iter()
-            .any(|b| matches!(b, Block::Paragraph { .. }));
-        let has_list = doc.blocks.iter().any(|b| matches!(b, Block::List { .. }));
-        let has_code = doc
-            .blocks
-            .iter()
-            .any(|b| matches!(b, Block::CodeBlock { .. }));
-
-        assert!(has_heading, "should have heading");
-        assert!(has_paragraph, "should have paragraph");
-        assert!(has_list, "should have list");
-        assert!(has_code, "should have code block");
-    }
-
-    #[test]
-    fn handles_empty_input() {
-        let parser = Parser::with_defaults("");
-        let doc = parser.parse().expect("should handle empty input");
-        assert_eq!(doc.blocks.len(), 0);
-    }
-
-    #[test]
-    fn handles_whitespace_only_input() {
-        let parser = Parser::with_defaults("   \n\t\n   ");
-        let doc = parser.parse().expect("should handle whitespace-only input");
-        // Should not create any blocks for whitespace-only input
-        assert_eq!(doc.blocks.len(), 0);
-    }
-
-    #[test]
-    fn parses_inline_elements() {
-        let input = "This has *emphasis* and `code` and [link](url).";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse inline elements");
-
-        assert_eq!(doc.blocks.len(), 1);
-        match &doc.blocks[0] {
-            Block::Paragraph { content, .. } => {
-                assert!(!content.is_empty());
-
-                // Check for different inline types
-                let has_text = content.iter().any(|i| matches!(i, Inline::Text(_)));
-                let _has_emphasis = content.iter().any(|i| matches!(i, Inline::Emphasis { .. }));
-                let _has_code = content.iter().any(|i| matches!(i, Inline::Code(_)));
-                let _has_link = content.iter().any(|i| matches!(i, Inline::Link { .. }));
-
-                assert!(has_text, "should have text");
-                // Note: The current implementation may not parse all inline elements perfectly
-                // This is expected as we're building incrementally
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn parses_nested_blockquotes() {
-        let input = "> Level 1\n>> Level 2\n>>> Level 3";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse nested blockquotes");
-
-        // Should create at least one blockquote
-        let blockquote_count = doc
-            .blocks
-            .iter()
-            .filter(|block| matches!(block, Block::BlockQuote { .. }))
-            .count();
-
-        assert!(blockquote_count >= 1, "should have at least one blockquote");
-    }
-
-    #[test]
-    fn handles_malformed_input_gracefully() {
-        let test_cases = vec![
-            "# Heading without content",
-            "```\nunclosed code block",
-            "> blockquote\nwithout continuation",
-            "- list item\nwithout proper spacing",
-        ];
-
-        for input in test_cases {
-            let parser = Parser::with_defaults(input);
-            let result = parser.parse();
-
-            // Should not panic and should produce some result
-            assert!(result.is_ok(), "should handle malformed input: {}", input);
-        }
-    }
-
-    #[test]
-    fn preserves_position_information() {
-        let input = "# Heading\n\nParagraph";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse with positions");
-
-        // Check that blocks have position information when available
-        for block in &doc.blocks {
-            match block {
-                Block::Heading { position, .. }
-                | Block::Paragraph { position, .. }
-                | Block::CodeBlock { position, .. }
-                | Block::BlockQuote { position, .. }
-                | Block::List { position, .. }
-                | Block::ThematicBreak { position, .. }
-                | Block::HtmlBlock { position, .. } => {
-                    // Position should be Some for blocks parsed by nom parser
-                    assert!(position.is_some(), "block should have position information");
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_parser_configuration() {
-        let input = "# Test";
-
-        // Test with strict mode
-        let strict_config = ParserConfig {
-            strict_commonmark: true,
-            ..ParserConfig::default()
-        };
-        let parser = Parser::new(input, strict_config);
-        assert!(parser.config().strict_commonmark);
-
-        // Test with non-strict mode
-        let non_strict_config = ParserConfig {
-            strict_commonmark: false,
-            ..ParserConfig::default()
-        };
-        let parser = Parser::new(input, non_strict_config);
-        assert!(!parser.config().strict_commonmark);
-    }
-
-    #[test]
-    fn test_error_recovery() {
-        let input = "# Valid heading\n\nSome text";
-        let config = ParserConfig {
-            strict_commonmark: false,
-            max_errors: Some(10),
-            ..ParserConfig::default()
-        };
-
-        let parser = Parser::new(input, config);
-        let result = parser.parse();
-
-        // Should succeed even with potential errors due to non-strict mode
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_unicode_handling() {
-        let input = "# 🌍 Unicode Heading\n\nText with émojis 🎉 and àccénts.";
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should handle Unicode");
-
-        assert_eq!(doc.blocks.len(), 2);
-
-        // Check that Unicode content is preserved
-        match &doc.blocks[0] {
-            Block::Heading { content, .. } => {
-                if let Inline::Text(text) = &content[0] {
-                    assert!(text.contains("🌍"));
-                }
-            }
-            _ => panic!("expected heading"),
-        }
-    }
-
-    #[test]
-    fn test_complex_document_structure() {
-        let input = r#"# Main Title
-
-This is an introduction paragraph.
-
-## Section 1
-
-> This is a blockquote with some content.
-> It spans multiple lines.
-
-### Subsection
-
-- First item
-- Second item
-  - Nested item
-- Third item
-
-```rust
-fn example() {
-    println!("Hello, world!");
-}
-```
-
-## Section 2
-
-Final paragraph with *emphasis* and `code`.
-
----
-
-End of document.
-"#;
-
-        let parser = Parser::with_defaults(input);
-        let doc = parser.parse().expect("should parse complex document");
-
-        // Should have multiple blocks
-        assert!(doc.blocks.len() >= 5);
-
-        // Check for variety of block types
-        let block_types: Vec<_> = doc
-            .blocks
-            .iter()
-            .map(|block| match block {
-                Block::Heading { .. } => "heading",
-                Block::Paragraph { .. } => "paragraph",
-                Block::CodeBlock { .. } => "code",
-                Block::BlockQuote { .. } => "blockquote",
-                Block::List { .. } => "list",
-                Block::ThematicBreak { .. } => "thematic_break",
-                Block::HtmlBlock { .. } => "html",
-            })
-            .collect();
-
-        assert!(block_types.contains(&"heading"));
-        assert!(block_types.contains(&"paragraph"));
-    }
-
-    #[test]
-    fn test_paragraph_builder() {
-        let mut builder = ParagraphBuilder::new();
-
-        // Test basic text accumulation
-        builder.push_text("Hello");
-        builder.push_text("world");
-
-        let mut blocks = Vec::new();
-        builder.flush_into(&mut blocks);
-
-        assert_eq!(blocks.len(), 1);
-        match &blocks[0] {
-            Block::Paragraph { content, .. } => {
-                assert_eq!(content.len(), 1);
-                if let Inline::Text(text) = &content[0] {
-                    assert!(text.contains("Hello"));
-                    assert!(text.contains("world"));
-                }
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
-
-    #[test]
-    fn test_inline_element_parsing() {
-        let mut builder = ParagraphBuilder::new();
-
-        // Test inline element accumulation
-        builder.push_inline(Inline::Text("Start ".to_string()));
-        builder.push_inline(Inline::Emphasis {
-            strong: false,
-            content: vec![Inline::Text("emphasized".to_string())],
-        });
-        builder.push_inline(Inline::Text(" end".to_string()));
-
-        let mut blocks = Vec::new();
-        builder.flush_into(&mut blocks);
-
-        assert_eq!(blocks.len(), 1);
-        match &blocks[0] {
-            Block::Paragraph { content, .. } => {
-                assert_eq!(content.len(), 3);
-                assert!(matches!(&content[0], Inline::Text(_)));
-                assert!(matches!(&content[1], Inline::Emphasis { .. }));
-                assert!(matches!(&content[2], Inline::Text(_)));
-            }
-            _ => panic!("expected paragraph"),
-        }
-    }
 }
